+++ apps/directus/directus-sync.config.local.js
@@ -1,0 +1,5 @@
+// ./directus-sync.config.js
+module.exports = {
+  extends: ['.directus-sync.config.base.js'],
+  directusUrl: 'http://cms.stack.localhost:8055',
+}