+++ libs/directus/directus-next/vite.config.ts
@@ -46,15 +46,16 @@ export default defineConfig({
           format: 'es',
           entryFileNames: '[name].mjs',
           preserveModules: false,
-          banner: chunk => chunk.name === 'server' ? '"use server";' : '',
+          banner: chunk => chunk.name === 'server' ? "import 'server-only';" : '',
         },
         {
           format: 'cjs',
           entryFileNames: '[name].js',
           preserveModules: false,
-          banner: chunk => chunk.name === 'server' ? '"use server";' : '',
+          banner: chunk => chunk.name === 'server' ? "require('server-only');" : '',
         },
       ],
+
       // External packages that should not be bundled into your library.
       external: id => isExternal(id) || id.startsWith('@okam/'),
     },