+++ libs/stack/stack-ui/src/components/Carousel/a11y/useCarouselA11yAnnouncer.tsx
@@ -7,7 +7,7 @@ import { useCarousel } from '../../../providers/Carousel'
 export function useCarouselA11yAnnouncer(props: TCarouselA11yAnnouncerProps) {
   const { controller, activeIndex, slides } = useCarousel()
   const { params } = controller ?? {}
-  const { a11y } = params ?? {}
+  const a11y = typeof params?.a11y === 'object' ? params.a11y : (params?.a11y === true ? { enabled: true } : undefined)
 
   return {
     a11y,