+++ libs/directus/directus-flexible-content/CHANGELOG.md
@@ -1,3 +1,16 @@
+## 2.0.2 (2026-03-17)
+
+### 🧱 Updated Dependencies
+
+- Updated directus-block to 2.0.2
+
+## 2.0.1 (2026-02-27)
+
+### 🧱 Updated Dependencies
+
+- Updated stack-ui to 2.0.1
+- Updated directus-block to 2.0.1
+
 # 2.0.0 (2026-02-05)
 
 ### 🧱 Updated Dependencies